make popup for second object
